Security contexts for workers
We have a need to provide a security context for workers to allow them to run as specific users. In particular we need them to be able to run as the same user as the coordinator as the shared image used in both needs a level of access to its filesystem. We are trying to run this cluster in an OpenShift environment, and have to deal with its restrictions on available user ids.

This change has been tested within our cluster, but I'm looking for feedback on whether this should be a separate security context for the workers vs coordinator.