Allow password authentication to be used with other authentication types
When a Trino cluster is configured to work with multiple authentication types through server.config.authenticationType key (e.g PASSWORD,OAUTH2), the password-authenticator.properties key is not configured in coordinator configmap because its template expects that authenticationType must equals to 'PASSWORD'. This prevents password authentication type to work and cluster coordinator to start.